[Remarks][NFC] Move the YAML serializer to its own header
llvm-svn: 366842
(cherry picked from commit cbbdc4183837dfea92cf840b93e0f12e2bcccffc)
